Split out `GlobalConfig` into its own header
This makes it easier to understand the reach of global variables /
global state in the config system.
